quality/test-recursively: Don't try to match 3rd party libs
With our new setup, having vendor-specific bundles, there are no third part
libraries. As such, we do not need to ignore errors from them, and that code can
be removed.

As it turns out, that code did not work too well, it made all failures appear
under the third party counter.

With the removal, that's no longer the case, and this fixes
keyboardio/Kaleidoscope-Bundle-Keyboardio#4.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>